+++ CHANGELOG.md
@@ -27,7 +27,12 @@ Table of Contents
 
 ### Bug Fixes
 
+* Fix a rare `java.nio.file.FileSystemAlreadyExistsException` when accessing filters concurrently.
+* Fix 1-click-maintenance not respecting the key of the `AttributeDifference`. 
+
 ### New Features
+
+* Metadata differences are now filterable via the recheck.ignore. A new `metadata.filter` has been added which can be imported with `import: metadata.filter` in `recheck.ignore`.
 
 ### Improvements
 
@@ -104,6 +109,7 @@ Table of Contents
 * You can now specify that a specific value-regex should be filtered, either globally, for specific elements or specific attributes. This allows to e.g. ignore a date, but still ensure that it is actually a valid date.
 * You can now specify to ignore a percentage color-diff, much like a pixel-diff. This allows to ignore, e.g. minor color changes or changes where opacity value is added or missing. Can be added globally, per element or attribute or combination thereof.
 * Add the option to filter elements based on whether they are inserted, deleted or changed.
+* The report output now displays the `retestId`: `input (input-retest-id) at 'html[1]/body[1]/input[1]'`. This may be used, for example, by a filter: `matcher: retestId=input-retest-id`.
 
 ### Improvements